requirements.txt
amqp==5.0.5
appdirs==1.4.4
apturl==0.5.2
asgiref==3.4.1
attrs==20.3.0
bcrypt==3.1.7
billiard==3.6.3.0
blinker==1.4
Brlapi==0.7.0
certifi==2020.12.5
cffi==1.14.3
chardet==4.0.0
chrome-gnome-shell==0.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
colorama==0.4.3
command-not-found==0.3
coverage==5.5
cryptography==3.2.1
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.1
distro==1.4.0
distro-info===0.23ubuntu1
Django==3.2.9
django-allauth==0.46.0
django-environ==0.8.1
djangorestframework==3.12.4
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
filelock==3.0.12
future==0.18.2
httplib2==0.14.0
idna==2.10
iniconfig==1.1.1
jwcrypto==0.8
kazam==1.4.5
keyring==18.0.1
kombu==5.0.2
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
lxml==4.6.2
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==1.1.1
mccabe==0.6.1
monotonic==1.5
netifaces==0.10.4
oauthlib==3.1.0
olefile==0.46
packaging==20.9
paramiko==2.6.0
pbr==5.5.1
pexpect==4.6.0
Pillow==7.0.0
pluggy==0.13.1
prompt-toolkit==3.0.18
protobuf==3.6.1
psycopg2==2.9.2
psycopg2-binary==2.9.2
py==1.10.0
pycairo==1.16.2
pycodestyle==2.7.0
pycparser==2.20
pycups==1.9.73
pyflakes==2.3.1
Pygments==2.3.1
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyodbc==4.0.22
pyOpenSSL==19.0.0
pyparsing==2.4.7
PyQt5==5.10.1
pyRFC3339==1.1
pytest==6.2.3
pytest-cov==2.11.1
python-apt==2.0.0+ubuntu0.20.4.6
python-dateutil==2.8.1
python-debian===0.1.36ubuntu1
python-editor==1.0.4
python-jwt==3.3.0
python3-openid==3.2.0
pytz==2021.1
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.25.1
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
simplejson==3.16.0
sip==4.19.21
six==1.15.0
SQLAlchemy==1.3.20
sqlparse==0.4.2
stevedore==3.3.0
systemd-python==234
toml==0.10.1
tornado==6.0.4
tornado-sqlalchemy==0.7.0
typecatcher==0.3
typing-extensions==3.7.4.3
ubuntu-advantage-tools==27.4
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.4
usb-creator==0.3.7
uvloop==0.14.0
vine==5.0.0
virtualenv==20.4.3
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
wadllib==1.3.3
wcwidth==0.2.5
Werkzeug==1.0.1
xkit==0.0.0
zipp==3.4.1
